Fix production build bugs.
