+++ src/environments/environment.prod.ts
@@ -1,4 +1,4 @@
-import { Environment } from './environment';
+import { Environment } from './environment.interface';
 
 export const environment: Environment = {
   production: true,
